requirements.txt
python == 3.10
numpy == 1.26.4
scikit-learn == 1.4.1post1
xgboost == 2.0.3
fastapi == 0.110.0
uvicorn == 0.29.0
pydantic == 2.6.4
dvc == 3.48.4
pandas == 2.2.1
ruff == 0.3.2
pyarrow == 15.0.1
fastparquet == 2024.2.0
tqdm == 4.66.2
juypterlab == 4.1.5
matplotlib == 3.8.3
seaborn == 0.13.2
dvclive == 3.45.0
ruamel-yaml == 0.18.6
# python-box == {version=="^5.0.0"extras == ["ruamel.yaml"]}
# requires == ["poetry-core"]
# build-backend == "poetry.core.masonry.api"